Move a few more things into legacy.go.
