+++ sublogger.go
@@ -71,68 +71,3 @@ func (logger SubLogger) isZero() bool {
 func (logger *SubLogger) SetLogLevel(level Level) {
 	logger.setLevel(level)
 }
-
-// TODO(ericsnow) Everything below here is unnecessary now and should
-// be deprecated.
-
-func (logger SubLogger) getModule() *module {
-	if logger.module == nil {
-		return newRootModule()
-	}
-	return logger.module
-}
-
-// LogLevel returns the configured min log level of the logger.
-//
-// This is strictly an alias for MinLogLevel, intended to align with
-// a more commonly used (but less specific) name.
-func (logger SubLogger) LogLevel() Level {
-	return logger.getModule().MinLogLevel()
-}
-
-// EffectiveLogLevel returns the effective min log level of
-// the receiver - that is, messages with a lesser severity
-// level will be discarded.
-//
-// If the log level of the receiver is unspecified,
-// it will be taken from the effective log level of its
-// parent.
-func (logger SubLogger) EffectiveLogLevel() Level {
-	return EffectiveMinLevel(logger.getModule())
-}
-
-// IsLevelEnabled returns whether debugging is enabled
-// for the given log level.
-func (logger SubLogger) IsLevelEnabled(level Level) bool {
-	return IsLevelEnabled(logger.getModule(), level)
-}
-
-// IsErrorEnabled returns whether debugging is enabled
-// at error level.
-func (logger SubLogger) IsErrorEnabled() bool {
-	return logger.IsLevelEnabled(ERROR)
-}
-
-// IsWarningEnabled returns whether debugging is enabled
-// at warning level.
-func (logger SubLogger) IsWarningEnabled() bool {
-	return logger.IsLevelEnabled(WARNING)
-}
-
-// IsInfoEnabled returns whether debugging is enabled
-// at info level.
-func (logger SubLogger) IsInfoEnabled() bool {
-	return logger.IsLevelEnabled(INFO)
-}
-
-// IsDebugEnabled returns whether debugging is enabled
-// at debug level.
-func (logger SubLogger) IsDebugEnabled() bool {
-	return logger.IsLevelEnabled(DEBUG)
-}
-
-// IsTraceEnabled returns whether debugging is enabled
-// at trace level.
-func (logger SubLogger) IsTraceEnabled() bool {
-	return logger.IsLevelEnabled(TRACE)
-}
